Caterwauling gemini: move error-handlers to state dir
