+++ src/state/search.js
@@ -2,9 +2,11 @@
 import algoliasearch from 'algoliasearch/lite';
 import { useEffect, useReducer, useMemo } from 'react';
 import { mapValues, sumBy, pick } from 'lodash';
+
+import { allByKeys } from 'Shared/api';
+import useErrorHandlers from 'State/error-handlers';
+
 import { useAPI } from './api';
-import { allByKeys } from '../../shared/api';
-import useErrorHandlers from '../presenters/error-handlers';
 import starterKits from '../curated/starter-kits';
 
 // TODO: this is super hacky; this would probably work a lot better with algolia
